apps/auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Authentication microservice for Gaming Library project"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.11.7"}
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
sqlalchemy = "^2.0.41"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
pylint = "^3.3.7"
